add: cleanup program, works
