+++ program/src/main.rs
@@ -31,30 +31,36 @@ pub fn main() {
         execution_state_proof,
     } = serde_cbor::from_slice(&encoded_inputs).unwrap();
 
-    let mut is_valid = true;
     let prev_header: B256 = store.finalized_header.tree_hash_root();
     let prev_head = store.finalized_header.slot;
 
     // 1. Apply sync committee updates, if any
     for (index, update) in updates.iter().enumerate() {
-        println!("Processing update {} of {}", index + 1, updates.len());
-        is_valid = is_valid
-            && verify_update(update, expected_current_slot, &store, genesis_root, &forks).is_ok();
-        println!("Update is valid: {}", is_valid);
+        println!("Processing update {} of {}.", index + 1, updates.len());
+        let update_is_valid =
+            verify_update(update, expected_current_slot, &store, genesis_root, &forks).is_ok();
+
+        if !update_is_valid {
+            panic!("Update {} is invalid!", index + 1);
+        }
+        println!("Update {} is valid.", index + 1);
         apply_update(&mut store, update);
     }
 
     // 2. Apply finality update
-    is_valid = is_valid
-        && verify_finality_update(
-            &finality_update,
-            expected_current_slot,
-            &store,
-            genesis_root,
-            &forks,
-        )
-        .is_ok();
-    println!("Finality update is valid: {}", is_valid);
+    let finality_update_is_valid = verify_finality_update(
+        &finality_update,
+        expected_current_slot,
+        &store,
+        genesis_root,
+        &forks,
+    )
+    .is_ok();
+    if !finality_update_is_valid {
+        panic!("Finality update is invalid!");
+    }
+    println!("Finality update is valid.");
+
     apply_finality_update(&mut store, &finality_update);
 
     // 3. Verify execution state root proof
@@ -64,21 +70,19 @@ pub fn main() {
         .map(|b| Node::try_from(b.as_ref()).unwrap())
         .collect();
 
-    is_valid = is_valid
-        && is_valid_merkle_branch(
-            &Node::try_from(execution_state_proof.execution_state_root.as_ref()).unwrap(),
-            execution_state_branch_nodes.iter(),
-            MERKLE_BRANCH_DEPTH,
-            MERKLE_BRANCH_INDEX,
-            &Node::try_from(store.finalized_header.body_root.as_ref()).unwrap(),
-        );
-    println!("Execution state root proof is valid: {}", is_valid);
+    let execution_state_proof_is_valid = is_valid_merkle_branch(
+        &Node::try_from(execution_state_proof.execution_state_root.as_ref()).unwrap(),
+        execution_state_branch_nodes.iter(),
+        MERKLE_BRANCH_DEPTH,
+        MERKLE_BRANCH_INDEX,
+        &Node::try_from(store.finalized_header.body_root.as_ref()).unwrap(),
+    );
+    if !execution_state_proof_is_valid {
+        panic!("Execution state root proof is invalid!");
+    }
+    println!("Execution state root proof is valid.");
 
-    // 4. Asset all constraints are valid
-    assert!(is_valid);
-    println!("All constraints are valid");
-
-    // 5. Commit new state root, header, and sync committee for usage in the on-chain contract
+    // 4. Commit new state root, header, and sync committee for usage in the on-chain contract
     let header: B256 = store.finalized_header.tree_hash_root();
     let sync_committee_hash: B256 = store.current_sync_committee.tree_hash_root();
     let next_sync_committee_hash: B256 = match &mut store.next_sync_committee {
