+++ script/Cargo.toml
@@ -14,10 +14,6 @@ name = "operator"
 path = "./bin/operator.rs"
 
 [[bin]]
-name = "gen-inputs"
-path = "./bin/gen-inputs.rs"
-
-[[bin]]
 name = "test"
 path = "./bin/test.rs"
